+++ src/main/java/com/seeme/domain/weather/WeatherWeekResDto.java
@@ -7,12 +7,12 @@ import lombok.Getter;
 @Builder
 public class WeatherWeekResDto {
 	private final String day;
-	private final Double amRain;
+	private final int amRain;
 	private final int amRainPercent;
 	private final String amIcon;
-	private final Double pmRain;
+	private final int pmRain;
 	private final int pmRainPercent;
 	private final String pmIcon;
-	private final Double max;
-	private final Double min;
+	private final int max;
+	private final int min;
 }
